requirements.txt
yfinance>=0.2.66
psycopg[binary]>=3.2
python-dotenv>=1.0
